+++ external/app-main/src/main/java/com/pharmacy/app/main/PharmacyBackApplication.java
@@ -1,0 +1,19 @@
+package com.pharmacy.app.main;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@SpringBootApplication
+@ComponentScan(basePackages = {"com.pharmacy.*"})
+@EntityScan(basePackages = {"com.pharmacy.*"})
+@EnableJpaRepositories(basePackages = {"com.pharmacy.*"})
+public class PharmacyBackApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(PharmacyBackApplication.class, args);
+    }
+
+}
