+++ external/app-main/src/main/resources/application.properties
