+++ external/persistence/src/main/java/com/pharmacy/persistence/mappers/ProductMapperImpl.java
@@ -1,0 +1,21 @@
+package com.pharmacy.persistence.mappers;
+
+import com.pharmacy.application.contracts.dtos.ProductDTO;
+import com.pharmacy.domain.model.Product;
+import com.pharmacy.application.contracts.mappers.ProductMapper;
+import org.mapstruct.Mapper;
+import org.mapstruct.MappingConstants;
+import org.mapstruct.ReportingPolicy;
+
+import java.util.List;
+
+@Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE, componentModel = MappingConstants.ComponentModel.SPRING)
+public interface ProductMapperImpl extends ProductMapper {
+
+    @Override
+    ProductDTO toDto(Product product);
+
+    @Override
+    List<ProductDTO> toDto(List<Product> productList);
+
+}
