+++ external/webapi/src/main/java/com/pharmacy/webapi/exceptions/ExceptionResponse.java
@@ -1,0 +1,6 @@
+package com.pharmacy.webapi.exceptions;
+
+import org.springframework.http.HttpStatus;
+
+public record ExceptionResponse(String message, HttpStatus status) {
+}
