--- infrastructure/pom.xml
@@ -1,28 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <parent>
-        <artifactId>pharmacy-back</artifactId>
-        <groupId>com.pharmacy</groupId>
-        <version>0.0.1-SNAPSHOT</version>
-    </parent>
-
-    <modelVersion>4.0.0</modelVersion>
-    <groupId>com.pharmacy</groupId>
-    <artifactId>infrastructure</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>app-main</module>
-        <module>repository</module>
-    </modules>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.pharmacy</groupId>
-            <artifactId>application-contracts</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-        </dependency>
-    </dependencies>
-
-</project>
