--- presentation/rest/src/main/java/com/pharmacy/presentation/exceptions/ExceptionResponse.java
@@ -1,6 +1,0 @@
-package com.pharmacy.presentation.exceptions;
-
-import org.springframework.http.HttpStatus;
-
-public record ExceptionResponse(String message, HttpStatus status) {
-}
